Reorganized project and deleted used gui code
