+++ src/models/InjurySeverity.java
@@ -1,4 +1,4 @@
-package backend.mainController;
+package models;
 
 public enum InjurySeverity {
     LIFE_THREATENING, NON_LIFE_THREATENING
