+++ src/models/MapGrid.java
@@ -1,4 +1,4 @@
-package backend.mainController;
+package models;
 
 import java.util.LinkedList;
 import java.util.List;
