+++ src/models/tests/AmbulanceTest.java
@@ -1,7 +1,7 @@
-package backend.mainController.tests;
+package models.tests;
 
-import backend.mainController.Ambulance;
-import backend.mainController.Point;
+import models.Ambulance;
+import models.Point;
 import org.junit.jupiter.api.Test;
 
 import static org.junit.jupiter.api.Assertions.*;
